add config for test data
